+++ torchmetrics/functional/__init__.py
@@ -24,7 +24,7 @@ from torchmetrics.functional.classification.calibration_error import calibration
 from torchmetrics.functional.classification.cohen_kappa import cohen_kappa
 from torchmetrics.functional.classification.confusion_matrix import confusion_matrix
 from torchmetrics.functional.classification.dice import dice_score
-from torchmetrics.functional.classification.f_beta import f1, f1_score, fbeta
+from torchmetrics.functional.classification.f_beta import f1, f1_score, fbeta, fbeta_score
 from torchmetrics.functional.classification.hamming_distance import hamming_distance
 from torchmetrics.functional.classification.hinge import hinge, hinge_loss
 from torchmetrics.functional.classification.iou import iou  # noqa: F401
@@ -69,6 +69,7 @@ from torchmetrics.functional.text.bert import bert_score
 from torchmetrics.functional.text.bleu import bleu_score
 from torchmetrics.functional.text.cer import char_error_rate
 from torchmetrics.functional.text.chrf import chrf_score
+from torchmetrics.functional.text.eed import extended_edit_distance
 from torchmetrics.functional.text.mer import match_error_rate
 from torchmetrics.functional.text.rouge import rouge_score
 from torchmetrics.functional.text.sacre_bleu import sacre_bleu_score
@@ -93,9 +94,11 @@ __all__ = [
     "tweedie_deviance_score",
     "dice_score",
     "explained_variance",
+    "extended_edit_distance",
     "f1",
     "f1_score",
     "fbeta",
+    "fbeta_score",
     "hamming_distance",
     "hinge",
     "hinge_loss",