+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `SQuAD` ([#623](https://github.com/PyTorchLightning/metrics/pull/623))
   - `CHRFScore` ([#641](https://github.com/PyTorchLightning/metrics/pull/641))
   - `TranslationEditRate` ([#646](https://github.com/PyTorchLightning/metrics/pull/646))
+  - `ExtendedEditDistance` ([#668](https://github.com/PyTorchLightning/metrics/pull/668))
+
 
 - Added `MultiScaleSSIM` into image metrics ([#679](https://github.com/PyTorchLightning/metrics/pull/679))
 
@@ -53,6 +55,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed `torchmetrics.collections` to `torchmetrics.metrics_collections` to avoid clashing with system's `collections` package ([#695](https://github.com/PyTorchLightning/metrics/pull/695))
 
 
+- Changed dtype of metric state from `torch.float` to `torch.long` in `ConfusionMatrix` to accommodate larger values ([#708](https://github.com/PyTorchLightning/metrics/issues/708))
+
+
 ### Deprecated
 
 - Renamed IoU -> Jaccard Index ([#662](https://github.com/PyTorchLightning/metrics/pull/662))
@@ -87,6 +92,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed Hinge metric: ([#734](https://github.com/PyTorchLightning/metrics/pull/734))
   * `torchmetrics.functional.hinge` -> `torchmetrics.functional.hinge_loss`
   * `torchmetrics.Hinge` -> `torchmetrics.HingeLoss`
+
+- Renamed F-Beta metrics: ([#740](https://github.com/PyTorchLightning/metrics/pull/740))
+  * `torchmetrics.functional.fbeta` -> `torchmetrics.functional.fbeta_score`
+  * `torchmetrics.FBeta` -> `torchmetrics.FBetaScore`
 
 
 - Renamed pit metric: ([#737](https://github.com/PyTorchLightning/metrics/pull/737))